fix: user_id 타입 변경

+++ backend/src/main/java/com/a407/back/controller/basicController.java
@@ -1,21 +1,24 @@
 package com.a407.back.controller;
 
 import com.a407.back.dto.UserRequest;
+import com.a407.back.model.service.basicService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import com.a407.back.model.service.basicService;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 
 
 @RequiredArgsConstructor
 @RestController
 public class basicController {
+
     private final basicService basicService;
 
     @PostMapping("/post")
-    public ResponseEntity<Integer> postMethod(@RequestBody UserRequest user) {
-        int id = basicService.save(user.toEntity());
+    public ResponseEntity<Long> postMethod(@RequestBody UserRequest user) {
+        Long id = basicService.save(user.toEntity());
         return ResponseEntity.status(HttpStatus.CREATED).body(id);
     }
 }
