+++ backend/src/main/java/com/a407/back/dto/UserRequest.java
@@ -4,13 +4,13 @@ import com.a407.back.domain.User;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 public class UserRequest {
-    private int id;
+
+    private Long id;
     private String password;
     private String email;
