+++ backend/src/main/java/com/a407/back/model/service/basicService.java
@@ -6,14 +6,14 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.Optional;
-
 @Service
 @RequiredArgsConstructor
 public class basicService {
+
     private final basicRepository basicRepository;
+
     @Transactional
-    public int save(User user) {
+    public Long save(User user) {
         return basicRepository.save(user).getId();
     }
 }
